Implement Material constructor and add get_material_name function in bindings
